perf(output): Replace Handlebars with direct string builders for output generation
Eliminate the ~140ms Handlebars module import + template compilation that
blocked the output generation critical path. Replace with zero-overhead
direct string builder functions for XML, Markdown, and Plain styles.

Why: Handlebars (350KB of JavaScript) takes ~140ms to parse, compile, and
initialize on every cold CLI invocation. This cost sat on the output
generation critical path, which is the longest-running phase of pack().
Direct string building using array push + join produces identical output
with no module loading overhead.

Changes:
- Add buildXmlOutput, buildMarkdownOutput, buildPlainOutput functions
  that directly construct output strings without template engines
- Extract extensionToLanguageMap to outputFileLanguage.ts (Handlebars-free)
  so the markdown builder can resolve language hints without importing
  Handlebars through outputStyleUtils.ts
- Keep Handlebars templates intact for skill generation path and tests
- Update test mocks from generateHandlebarOutput to buildDirectOutput

Benchmark (1000 files, cold-start, 5 runs trimmed mean):
  Before: ~1368ms (median 1362ms, range 1319–1413ms)
  After:  ~1249ms (median 1243ms, range 1237–1278ms)
  Improvement: ~120ms = 8.7% of total CLI wall time
