+++ tests/core/output/outputGenerate.test.ts
@@ -16,7 +16,7 @@ const createStrictXmlParser = () => {
 describe('outputGenerate', () => {
   const mockDeps = {
     buildOutputGeneratorContext: vi.fn(),
-    generateHandlebarOutput: vi.fn(),
+    buildDirectOutput: vi.fn(),
     generateParsableXmlOutput: vi.fn(),
     generateParsableJsonOutput: vi.fn(),
     sortOutputFiles: vi.fn(),
@@ -47,7 +47,7 @@ describe('outputGenerate', () => {
       instruction: '',
       filesEnabled: true,
     });
-    mockDeps.generateHandlebarOutput.mockResolvedValue('mock output');
+    mockDeps.buildDirectOutput.mockReturnValue('mock output');
 
     const output = await generateOutput(
       [process.cwd()],
