+++ tests/core/output/outputGenerateDiffs.test.ts
@@ -59,7 +59,7 @@ describe('Output Generation with Diffs', () => {
       instruction: '',
       gitDiffResult,
     })),
-    generateHandlebarOutput: vi.fn(),
+    buildDirectOutput: vi.fn(),
     generateParsableXmlOutput: vi.fn(),
     generateParsableJsonOutput: vi.fn(),
     sortOutputFiles: vi.fn().mockResolvedValue(mockProcessedFiles),
@@ -70,8 +70,8 @@ describe('Output Generation with Diffs', () => {
     mockConfig.output.style = 'xml';
     mockConfig.output.parsableStyle = false;
 
-    // Mock the Handlebars output function to check for diffs in the template
-    mockDeps.generateHandlebarOutput.mockImplementation((_config, renderContext: RenderContext, _processedFiles) => {
+    // Mock the direct string builder function to check for diffs in the template
+    mockDeps.buildDirectOutput.mockImplementation((_style, renderContext: RenderContext, _processedFiles) => {
       // Verify that the renderContext has the gitDiffs property
       expect(renderContext.gitDiffWorkTree).toBe(sampleDiff);
 
@@ -97,8 +97,8 @@ describe('Output Generation with Diffs', () => {
     expect(output).toContain(sampleDiff);
     expect(output).toContain('</diffs>');
 
-    // Verify that the generateHandlebarOutput function was called
-    expect(mockDeps.generateHandlebarOutput).toHaveBeenCalled();
+    // Verify that the buildDirectOutput function was called
+    expect(mockDeps.buildDirectOutput).toHaveBeenCalled();
   });
 
   test('XML style output with parsableStyle should include diffs section', async () => {
@@ -142,8 +142,8 @@ describe('Output Generation with Diffs', () => {
     mockConfig.output.style = 'markdown';
     mockConfig.output.parsableStyle = false;
 
-    // Mock the Handlebars output function for markdown
-    mockDeps.generateHandlebarOutput.mockImplementation((_config, renderContext: RenderContext, _processedFiles) => {
+    // Mock the direct string builder function for markdown
+    mockDeps.buildDirectOutput.mockImplementation((_style, renderContext: RenderContext, _processedFiles) => {
       // Verify that the renderContext has the gitDiffs property
       expect(renderContext.gitDiffWorkTree).toBe(sampleDiff);
 
@@ -170,8 +170,8 @@ describe('Output Generation with Diffs', () => {
     expect(output).toContain(sampleDiff);
     expect(output).toContain('```');
 
-    // Verify that the generateHandlebarOutput function was called
-    expect(mockDeps.generateHandlebarOutput).toHaveBeenCalled();
+    // Verify that the buildDirectOutput function was called
+    expect(mockDeps.buildDirectOutput).toHaveBeenCalled();
   });
 
   test('Plain style output should include diffs section when includeDiffs is enabled', async () => {
@@ -179,8 +179,8 @@ describe('Output Generation with Diffs', () => {
     mockConfig.output.style = 'plain';
     mockConfig.output.parsableStyle = false;
 
-    // Mock the Handlebars output function for plain text
-    mockDeps.generateHandlebarOutput.mockImplementation((_config, renderContext: RenderContext, _processedFiles) => {
+    // Mock the direct string builder function for plain text
+    mockDeps.buildDirectOutput.mockImplementation((_style, renderContext: RenderContext, _processedFiles) => {
       expect(renderContext.gitDiffWorkTree).toBe(sampleDiff);
 
       // Simulate the plain text output
@@ -204,8 +204,8 @@ describe('Output Generation with Diffs', () => {
     expect(output).toContain('===============\nGit Diffs\n===============');
     expect(output).toContain(sampleDiff);
 
-    // Verify that the generateHandlebarOutput function was called
-    expect(mockDeps.generateHandlebarOutput).toHaveBeenCalled();
+    // Verify that the buildDirectOutput function was called
+    expect(mockDeps.buildDirectOutput).toHaveBeenCalled();
   });
 
   test('Output should not include diffs section when includeDiffs is disabled', async () => {
@@ -224,8 +224,8 @@ describe('Output Generation with Diffs', () => {
       // No gitDiffs property
     }));
 
-    // Mock the Handlebars output function
-    mockDeps.generateHandlebarOutput.mockImplementation((_config, renderContext: RenderContext, _processedFiles) => {
+    // Mock the direct string builder function
+    mockDeps.buildDirectOutput.mockImplementation((_style, renderContext: RenderContext, _processedFiles) => {
       // Verify that the renderContext does not have the gitDiffs property
       expect(renderContext.gitDiffWorkTree).toBeUndefined();
 
@@ -250,7 +250,7 @@ describe('Output Generation with Diffs', () => {
     expect(output).not.toContain('Git Diffs');
     expect(output).not.toContain(sampleDiff);
 
-    // Verify that the generateHandlebarOutput function was called
-    expect(mockDeps.generateHandlebarOutput).toHaveBeenCalled();
+    // Verify that the buildDirectOutput function was called
+    expect(mockDeps.buildDirectOutput).toHaveBeenCalled();
   });
 });
